Make functions, which helps to output to file and stdout
Factor the code (very marginally...) into functions. At least this way
we avoid a need to use either mkfifo or process subsitution to
accomplish the goal of outputting to a log file and stdout
simulatenously. It also is a first step to make the code more
maintainable. On the other hand, it screws up the git history a bit
for now.